guardando carrito

+++ Carrito/src/main/java/com/informatorio/Carrito/Controller/UsuarioController.java
@@ -1,12 +1,15 @@
 package com.informatorio.Carrito.Controller;
 
 
+import com.informatorio.Carrito.Entity.Producto;
 import com.informatorio.Carrito.Entity.Usuario;
 import com.informatorio.Carrito.Repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 public class UsuarioController {
@@ -38,6 +41,24 @@ public class UsuarioController {
 
     @DeleteMapping(value = "/usuario/{id}")
     public void borrarPorId(@PathVariable("id") Long id) {
-        usuarioRepository.deleteById(id);
+
+        Usuario usuario = usuarioRepository.findById(id).get();
+        if(!usuario.getBaja()) {
+            if (usuario.getCarritos().size() == 0) {
+                usuarioRepository.deleteById(id);
+            } else {
+                usuario.setBaja(true);
+            }
+        }
+    }
+
+    @GetMapping(value = "/usuario/ciudad/{ciudad}")
+    public List<Usuario> getUsuarioCiudad(@PathVariable("ciudad") String ciudad) {
+        return usuarioRepository.findByCiudad(ciudad);
+    }
+
+    @GetMapping(value = "/usuario/fecha/{fecha}")
+    public List<Usuario> getUsuarioFecha(@PathVariable("fecha") String fecha) {
+        return usuarioRepository.findAllByFechaCreacion(fecha);
     }
 }
